Enhance vertex handling and debugging in graph structure
- Updated multiple binary files: `cgdb`, `data.bin`, `free_space.bin`, and `map.bin` to reflect recent changes.
- Refactored vertex-related functions in `circle.c`, `vertex.c`, and `command_handler.c` to improve consistency and error handling.
- Added debugging print statements in `get_vertex_position`, `delete_vertex`, and other functions to aid in tracing execution flow and identifying issues.
- Updated function signatures to use `unsigned int` and `ushort` for vertex and channel indices, enhancing type safety.
- Enhanced CLI command handling to provide clearer feedback on vertex operations, including deletion and validation.
- Improved documentation in `cli-interface.md` and `graph-based-database.md` to reflect new functionalities and command usage.

+++ src/Graph_structure/circle.h
@@ -11,15 +11,15 @@ typedef struct {
 } CircleInfo;
 
 // Check if there is a circle starting from given vertex/channel/axis
-bool has_circle(int vertex_index, int channel_index, int axis_number);
+bool has_circle(unsigned int vertex_index, ushort channel_index, ushort axis_number);
 
 // Get detailed information about the circle
-CircleInfo* get_circle_info(int vertex_index, int channel_index, int axis_number);
+CircleInfo* get_circle_info(unsigned int vertex_index, ushort channel_index, ushort axis_number);
 
 // Free circle info structure
 void free_circle_info(CircleInfo* info);
 
 // Add this function declaration
-bool is_in_garbage_circle(uint vertex_index);
+bool is_in_garbage_circle(unsigned int vertex_index);
 
 #endif // CIRCLE_H
